+++ src/frontend/ui/comparaison/useComparaison.tsx
@@ -29,6 +29,7 @@ import { ContenuFondDeRoulementNetGlobal } from "../établissement-territorial-m
 import { ContenuRésultatNetComptable } from "../établissement-territorial-médico-social/InfoBulle/ContenuRésultatNetComptable";
 import { ContenuTauxDeVétustéConstruction } from "../établissement-territorial-médico-social/InfoBulle/ContenuTauxDeVétustéConstruction";
 import { ContenuDuTauxOccupationESMS } from "../établissement-territorial-médico-social/InfoBulle/ContenuTauxOccupationESMS";
+import { ContenuTauxOccupationGlobal } from "../établissement-territorial-médico-social/InfoBulle/ContenuTauxOccupationGlobal";
 import { ContenuTauxRéalisationActivité } from "../établissement-territorial-médico-social/InfoBulle/ContenuTauxRéalisationActivité";
 import { ContenuNombreDeJournéesPSYetSSR } from "../établissement-territorial-sanitaire/InfoBulle/ContenuNombreDeJournéesPSYetSSR";
 import { ContenuNombreDeSéjourMCO } from "../établissement-territorial-sanitaire/InfoBulle/ContenuNombreDeSéjourMCO";
@@ -201,10 +202,14 @@ export function useComparaison() {
       case "internat":
       case "autres":
       case "seances":
-      case "global":
         return {
           contenu: <ContenuDuTauxOccupationESMS dateDeMiseÀJour={StringFormater.formatDate(dates.date_mis_a_jour_cnsa)} source={wording.CNSA} />,
           titre: getTitreModalESMS(name),
+        };
+      case "global":
+        return {
+          contenu: <ContenuTauxOccupationGlobal dateDeMiseÀJour={StringFormater.formatDate(dates.date_mis_a_jour_cnsa)} source={wording.CNSA} />,
+          titre: wording.TAUX_OCCUPATION_GLOBAL,
         };
       case "prestationExterne":
         return {
@@ -350,8 +355,6 @@ export function useComparaison() {
         return wording.TAUX_OCCUPATION_INTERNAT;
       case "autres":
         return wording.TAUX_OCCUPATION_AUTRE;
-      case "global":
-        return wording.TAUX_OCCUPATION_GLOBAL;
       default:
         return wording.TAUX_OCCUPATION_SEANCES
     }
@@ -384,7 +387,7 @@ export function useComparaison() {
         { label: "TO Internat", nomComplet: "Taux d’occupation internat", key: "internat", info: true, sort: true, orderBy: "taux_occupation_internat" },
         { label: "TO Autre 1, 2 et 3", nomComplet: "Taux d’occupation autre 1, 2 et 3", key: "autres", info: true, sort: true, orderBy: "taux_occupation_autres" },
         { label: "TO Séances", nomComplet: "Taux d'occupation Séances", key: "seances", info: true, sort: true, orderBy: "taux_occupation_seances" },
-        { label: "TO global occupation", nomComplet: "Taux d'occupation global", key: "global", info: true, sort: true, orderBy: "taux_occupation_global" },
+        { label: "TO global", nomComplet: "Taux d'occupation global", key: "global", info: true, sort: true, orderBy: "taux_occupation_global" },
         { label: "Tx de prest ext sur les prest directes", nomComplet: "Taux de prestations externes sur les prestations directes", key: "prestationExterne", info: true, sort: true, orderBy: "taux_prestation_externes" },
         { label: "Tx de rotation du personnel sur effectifs réels", nomComplet: "Taux de rotation du personnel sur effectifs réels", key: "rotationPersonnel", info: true, sort: true, orderBy: "taux_rotation_personnel" },
         { label: "Tx d'ETP vacants au 31/12", nomComplet: "Taux d'ETP vacants au 31/12", key: "etpVacant", info: true, sort: true, orderBy: "taux_etp_vacants" },