+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Bug Fixes
 - hel-971: [Gestion des listes] Excel intitulet de colonne à revoir
 
+
+- hel-969: Comparaison Colonne Taux global d'occupation
 - hel-968: Fix du bug de depliement/repliement des blocs dans les fiches d’établissements
 
 ### Features